refacotring, moved all security contexts in one folder
