+++ src/main/java/com/github/polimi_mt_acg/back2school/api/v1/security_contexts/ParentOfStudentSecurityContext.java
@@ -1,6 +1,7 @@
-package com.github.polimi_mt_acg.back2school.api.v1.parents;
+package com.github.polimi_mt_acg.back2school.api.v1.security_contexts;
 
 import com.github.polimi_mt_acg.back2school.api.v1.security_contexts.InvalidTemplateParameterException;
+import com.github.polimi_mt_acg.back2school.api.v1.security_contexts.ParentOfStudentSecured;
 import com.github.polimi_mt_acg.back2school.api.v1.security_contexts.SecurityContextPriority;
 import com.github.polimi_mt_acg.back2school.model.AuthenticationSession;
 import com.github.polimi_mt_acg.back2school.model.User;
@@ -19,61 +20,65 @@ import javax.ws.rs.core.UriInfo;
 import javax.ws.rs.ext.Provider;
 import org.hibernate.Session;
 
-@SameParentSecured
+@ParentOfStudentSecured
 @Provider
-@Priority(SecurityContextPriority.SAME_PARENT)
-public class SameParentSecurityContext implements ContainerRequestFilter {
+@Priority(SecurityContextPriority.PARENT_OF_STUDENT)
+public class ParentOfStudentSecurityContext implements ContainerRequestFilter {
 
   @Override
   public void filter(ContainerRequestContext requestContext) throws IOException {
     DatabaseHandler dbi = DatabaseHandler.getInstance();
     Session session = dbi.getNewSession();
     session.beginTransaction();
-    User currentUser = AuthenticationSession.getCurrentUser(requestContext, session);
+    User parent = AuthenticationSession.getCurrentUser(requestContext, session);
 
     // Abort the filter chain with a 401 status code response
 
     // if there is no user logged in
-    if (currentUser == null) {
+    if (parent == null) {
       requestContext.abortWith(Response.status(Response.Status.UNAUTHORIZED).build());
       return;
     }
 
     // if the user logged has not the correct role
-    if (currentUser.getRole() != Role.PARENT) {
+    if (parent.getRole() != Role.PARENT) {
       return; // Move control to next filter
     }
 
-    // Get the parent id template parameter
+    // Get the student id template parameter
     UriInfo uriInfo = requestContext.getUriInfo();
     MultivaluedMap<String, String> pathParameter = uriInfo.getPathParameters();
 
     if (!pathParameter.containsKey("id")) {
       throw new InvalidTemplateParameterException(
-              "Could not find 'id' template parameter in URI. Maybe you annotated a non /parents/{id}/* REST endpoint?");
+          "Could not find 'id' template parameter in URI. Maybe you annotated a non /students/{id}/* REST endpoint?");
     }
 
-    // Get parent id from request's URI
-    int parentId = Integer.parseInt(pathParameter.getFirst("id"));
+    // Get student id from request's URI
+    int studentId = Integer.parseInt(pathParameter.getFirst("id"));
     List<User> result =
-            DatabaseHandler.getInstance()
-                    .getListSelectFromWhereEqual(User.class, User_.id, parentId, session);
+        DatabaseHandler.getInstance()
+            .getListSelectFromWhereEqual(User.class, User_.id, studentId, session);
     if (result.isEmpty()) {
       requestContext.abortWith(Response.status(Status.NOT_FOUND).build());
     }
 
-    // Check is requested parent is the same requesting parent
-    boolean isSameParent = false;
-    if (currentUser.getId() == result.get(0).getId())
-    {
-      isSameParent = true;
+    // Check is requested student is child of requesting parent
+    User student = result.get(0);
+    List<User> children = parent.getChildren();
+    boolean isParent = false;
+    for (User child : children) {
+      if (child.getId() == student.getId()) {
+        isParent = true;
+        break;
+      }
     }
 
     session.getTransaction().commit();
     session.close();
 
-    if (!isSameParent) {
-      requestContext.abortWith(Response.status(Response.Status.FORBIDDEN).build());
+    if (!isParent) {
+      requestContext.abortWith(Response.status(Response.Status.UNAUTHORIZED).build());
     }
   }
 }
