+++ src/main/java/com/github/polimi_mt_acg/back2school/api/v1/security_contexts/SameParentSecurityContext.java
@@ -1,6 +1,7 @@
-package com.github.polimi_mt_acg.back2school.api.v1.teachers;
+package com.github.polimi_mt_acg.back2school.api.v1.security_contexts;
 
 import com.github.polimi_mt_acg.back2school.api.v1.security_contexts.InvalidTemplateParameterException;
+import com.github.polimi_mt_acg.back2school.api.v1.security_contexts.SameParentSecured;
 import com.github.polimi_mt_acg.back2school.api.v1.security_contexts.SecurityContextPriority;
 import com.github.polimi_mt_acg.back2school.model.AuthenticationSession;
 import com.github.polimi_mt_acg.back2school.model.User;
@@ -9,7 +10,6 @@ import com.github.polimi_mt_acg.back2school.model.User_;
 import com.github.polimi_mt_acg.back2school.utils.DatabaseHandler;
 import java.io.IOException;
 import java.util.List;
-import java.util.Optional;
 import javax.annotation.Priority;
 import javax.ws.rs.container.ContainerRequestContext;
 import javax.ws.rs.container.ContainerRequestFilter;
@@ -20,14 +20,17 @@ import javax.ws.rs.core.UriInfo;
 import javax.ws.rs.ext.Provider;
 import org.hibernate.Session;
 
-@SameTeacherSecured
+@SameParentSecured
 @Provider
-@Priority(SecurityContextPriority.SAME_TEACHER)
-public class SameTeacherSecurityContext implements ContainerRequestFilter {
+@Priority(SecurityContextPriority.SAME_PARENT)
+public class SameParentSecurityContext implements ContainerRequestFilter {
 
   @Override
   public void filter(ContainerRequestContext requestContext) throws IOException {
-    User currentUser = AuthenticationSession.getCurrentUser(requestContext);
+    DatabaseHandler dbi = DatabaseHandler.getInstance();
+    Session session = dbi.getNewSession();
+    session.beginTransaction();
+    User currentUser = AuthenticationSession.getCurrentUser(requestContext, session);
 
     // Abort the filter chain with a 401 status code response
 
@@ -38,27 +41,39 @@ public class SameTeacherSecurityContext implements ContainerRequestFilter {
     }
 
     // if the user logged has not the correct role
-    if (currentUser.getRole() != Role.TEACHER) {
+    if (currentUser.getRole() != Role.PARENT) {
       return; // Move control to next filter
     }
 
-    // Get the teacher id template parameter
+    // Get the parent id template parameter
     UriInfo uriInfo = requestContext.getUriInfo();
     MultivaluedMap<String, String> pathParameter = uriInfo.getPathParameters();
 
     if (!pathParameter.containsKey("id")) {
       throw new InvalidTemplateParameterException(
-          "Could not find 'id' template parameter in URI. Maybe you annotated a non /teachers/{id}/* REST endpoint?");
+              "Could not find 'id' template parameter in URI. Maybe you annotated a non /parents/{id}/* REST endpoint?");
     }
 
-    // Get teacher id from request's URI
-    int teacherId = Integer.parseInt(pathParameter.getFirst("id"));
-    Optional<User> teacherOpt = DatabaseHandler.fetchEntityBy(User.class, User_.id, teacherId);
-    if (!teacherOpt.isPresent()) {
+    // Get parent id from request's URI
+    int parentId = Integer.parseInt(pathParameter.getFirst("id"));
+    List<User> result =
+            DatabaseHandler.getInstance()
+                    .getListSelectFromWhereEqual(User.class, User_.id, parentId, session);
+    if (result.isEmpty()) {
       requestContext.abortWith(Response.status(Status.NOT_FOUND).build());
     }
 
-    if (currentUser.getId() != teacherOpt.get().getId()) {
+    // Check is requested parent is the same requesting parent
+    boolean isSameParent = false;
+    if (currentUser.getId() == result.get(0).getId())
+    {
+      isSameParent = true;
+    }
+
+    session.getTransaction().commit();
+    session.close();
+
+    if (!isSameParent) {
       requestContext.abortWith(Response.status(Response.Status.FORBIDDEN).build());
     }
   }
