+++ src/main/java/com/github/polimi_mt_acg/back2school/api/v1/security_contexts/TeacherOfStudentSecured.java
@@ -1,4 +1,4 @@
-package com.github.polimi_mt_acg.back2school.api.v1.students;
+package com.github.polimi_mt_acg.back2school.api.v1.security_contexts;
 
 import static java.lang.annotation.ElementType.METHOD;
 import static java.lang.annotation.ElementType.TYPE;
